requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
appdirs==1.4.4
astroid==3.2.2
attrs==23.2.0
beautifulsoup4==4.13.0b2
bs4==0.0.2
certifi==2024.2.2
cffi==1.17.0rc1
cfgv==3.4.0
charset-normalizer==3.3.2
colorama==0.4.6
cssselect==1.2.0
dill==0.3.8
distlib==0.3.8
fake-useragent==1.5.1
filelock==3.14.0
greenlet==3.0.3
h11==0.14.0
identify==2.5.36
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
isort==6.0.0b2
lxml==5.2.2
mccabe==0.7.0
nodeenv==1.8.0
outcome==1.3.0.post0
packaging==24.0
parse==1.20.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
pycparser==2.22
pyee==11.1.0
pylint==3.2.2
pylint-pytest==2.0.0a0
PyMySQL==1.1.1
pyppeteer==2.0.0
pyquery==2.0.0
PySocks==1.7.1
pytest==8.2.1
pytest-variables==3.1.0
PyYAML==6.0.1
requests==2.32.2
requests-html==0.10.0
selenium==4.21.0
setuptools==70.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.30
toml==0.10.2
tomlkit==0.12.5
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.0
urllib3==2.2.1
virtualenv==20.26.2
w3lib==2.1.2
websockets==12.0
wsproto==1.2.0
zipp==3.19.0
